Isolate GR00T N1.7 model changes
